Modify RealtimeMarkingScheme for moving to OMR
